lib: dk_buttons_and_leds: add shell for buttons
There are applications where it would be convinient,
especially during testing to be able to simulate button press.

The dk_buttons_and_leds module seems to be a good place,
to inject simulated button presses for the application above.

This addition is disabled by default, and can be enabled through Kconfg.

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>