tests: samples: Add platform_allow section to configs without it
Entries in platform_allow are copies of integration_platforms.
This change is needed for dynamic scope optimization for twister.
A test_plan.py script selects which configurations to add to
the scope. If a test is modified, it is scoped for all applicable
platforms. This means for cases in the commit, selecting >600
platforms available in the upstream and nrf. Platform_allow
works as a filter limiting the scope.

Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>