samples: matter: Repair bindings via uart in Light Switch Sample
After the unification of the matter-binding mechanism in the
BindingData struct there is an additional callback for
invoking the command function.
It must be assigned because it is used within the
BindingManager module.

Moreover, the RunBoundClusterAction method must check whether
the InvokeCommandFunc is not NULL and does not allow doing any
binding action otherwise..

The InvokeCommandFunc callback was not being assigned when
we used UART CLI to send on/off command requests which
cused a crash.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>