Bluetooth: Controller: New default event length and spacing for ISO
When ISO is enabled, we want to set asside as much time as possible
for ISO activities.

By setting BT_CTLR_SDC_MAX_CONN_EVENT_LEN_DEFAULT=0, we ensure that
the ACL will utilize as little time as possible and avoid increasing
the data length capabilities.

The BT_CTLR_SDC_CENTRAL_ACL_EVENT_SPACING_DEFAULT needs to be updated
accordingly - we want the ACLs to be spaced apart to avoid scheduling
conflicts. By selecting the value of 30 ms, we ensure that:
 - The ACLs are non-overlapping given that they are using an interval
   which is greater than 30 ms.
 - The ACL events are not overlapping with ISO events, given that the
   ISO interval is a multiple of 7.5 ms or 10 ms.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>