net: lwm2m_client_utils: Use DTLS_CID_SUPPORTED
Use SUPPORTED instead of ENABLED for the Connection
Identifier settings.

This causes modem to send zero length CID to server,
indicating that we support CID but don't require one
for the reponse packets.
If server supports CID, it will generate one for us
that we use for uplink packets.

There is really no benefit for using CID on both
direction. Using it uplink direction is enough.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>