samples: matter: replace std::function with custom task
std::function<void()> is currently used as a task that is to
be executed in the application thread, but this class is not
trivially copyable so passing it over Zephyr's msgq may lead
to resource leaks or other unexpected behavior in certain
circumstances (for example, when a functor that std::function
is initialized with is big enough that it cannot fit in the
std::function static storage and must be allocated on heap).

Replace std::function with a lightweight replacement that
ensures that it is only used with simple lambda functions
that can be safely serialized.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>