samples: cellular: modem_shell: connect UDP sockets
This commit updates the sample to also connect UDP
sockets. UDP sockets are connectionless, however
the modem library supports invoking `connect` on a UDP
socket to set its peer address.
This simplifies the use of send which will internally
use the peer address.
This also fixes a problem when setting `RAI_NO_DATA`
on a UDP socket, since `RAI_NO_DATA` requires the socket
to be connected.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>