applications: nrf5340_audio: Fix CIS TWS MIC stream disappeared
Fixed the issue that CIS gateway might not setup MIC stream properly.
There could be a chance the sink stream setup failed, but before the
retry, the source stream from same ISO setup properly which also
setup the sink. This cause the state out of sync for retrying, and
leads to retry mechanism failed.
OCT-NONE
Signed-off-by: Jui-Chou Chung <jui-chou.chung@nordicsemi.no>