nrf_security: Fixed compliance issues
These changes were done to fix compliance issues.

The changes come from a mix of running checkpatch --fix-inplace,
clang-format, and manually fixing checkpatch and license issues.

When checkpatch was run it appeared to make changes to code that was
already non-compliant.

It is not clear why we have non-compliant code in the repo. But I
decided to include the changes anyway since any patches to these files
will trigger a checkpatch failure.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>