scripts: add cert_tool.py
This is a script to generate elliptic curve private keys, create a
certificate signing request for a private key, create root CA and
subordinate CA certificates and sign a certificate signing request
with a CA.

The keys are not encrypted with a password or any other protection. It
is strictly for testing and evaluation purposes.

It is only tested with Azure IoT Hub and Azure DPS, but doesn't include
anything specific for Azure, so in theory it should work with other
cloud vendors.

usage: cert_tool.py [-h] {client_key,root_ca,sub_ca,csr,sign,sign_root}

All file-related arguments have default values. Client key/csr/cert is
stored in ./certs folder and CA keys/certs are stored in ./ca.

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>