samples: matter: Fix configuration scheme
In Matter samples, we have the wrong configuration schema
regarding pm_static files. We cannot have it in the
configuration directory and set the PM_STATIC_YML_FILE in
the CMakeLists because we do not allow users to overwrite it.

In this commit:
- Removed redundant prj_no_dfu.conf configuration.
- No-dfu variant is needed only for nrf21540 DK.
- moved all pm_static files to samples' main directories and
renamed them to be consistent with the NCS build system and
partition manager requirements.
- Updated documentation regarding the changes.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>