tfm: 54L: UART configuration
Use UART20 instead of UART22 for the tfm_ns board.

Use UART30 instead of 22 for the TF-M image.

This was done so the TF-M image and the non-secure image could use
different UARTS.

Also, instead of using the same pins (1.4 and 1.5) for NS and S, we
use

1.4, 1.5 for NS
and
0.0 and 0.1 for S

This aligns better with other platforms.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>