Bluetooth: Mesh: Make metadata script handle multiple comp data
This makes the metadata extraction script handle applications with
multiple instances of composition data, where the composition data to
use is picked at runtime.

When more than one instance is found, no encoded metadata is printed as
part of the build output, as users would need to look at the JSON
anyways to decide which composition data to use when doing the update.

Signed-off-by: Ludvig Jordet <ludvig.jordet@nordicsemi.no>