net: lwm2m_client_utils: Fix supported protocols list for FOTA
FOTA object has list of supported protocols.
We assumed that when CA chain is present in certain sec_tag,
we can use HTTPS or CoAPS. This is OK.
But AVSystem uses mutual DTLS authentication on CoAPS so
the sec_tag we use, might not use CA chain at all.
So when only PSK credentials are present, claim that we
support CoAPS but not HTTPS.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>