Bluetooth: Host: allow reducing initiator priority
By default, the initiator in the SoftDevice Controller operates at a
higher priority than other BT roles. This may result in unwanted
scheduling conflicts with other roles.
On the primary channel, the initiator can be scheduled such that
conflicts can be avoided, by choosing an interval that is equal or a
multiple of the interval of other roles. However, the secondary channel
timings are set by the peer device. This new host extension allows
to (temporarily) reduce the priority the initiator uses when following
AUX pointers.

Signed-off-by: Jan Müller <jan.mueller@nordicsemi.no>